Move functionality needed for API to aggregator_core crate.
I plan to implement the Data Plane API in a separate crate, which will
depend on the new janus_aggregator_core crate for Datastore interaction,
and which the janus_aggregator crate will (optionally) depend on to
expose the API based on a feature.

This is mostly straightforward, though I had to make a few interface
changes to decouple interactions that are now cross-crate. The most
controversial is probably that there are now multiple
{Uploadable,Aggregatable,Collectable}QueryType traits: the traits in
janus_aggregator_core are the "bulk" of the functionality in these
traits, the traits in janus_aggregator are solely the functionality that
needs access to aggregator::Error.

While I'm at it, clean up unnecessary dependencies as reported by
cargo-udeps.
