+++ aggregator/src/aggregator/aggregate_init_tests.rs
@@ -1,12 +1,12 @@
-use crate::{
-    aggregator::{aggregator_filter, tests::generate_helper_report_share, Config},
+use crate::aggregator::{aggregator_filter, tests::generate_helper_report_share, Config};
+use http::{header::CONTENT_TYPE, StatusCode};
+use janus_aggregator_core::{
     datastore::{
         test_util::{ephemeral_datastore, EphemeralDatastore},
         Datastore,
     },
     task::{test_util::TaskBuilder, QueryType, Task},
 };
-use http::{header::CONTENT_TYPE, StatusCode};
 use janus_core::{
     task::VdafInstance,
     test_util::{dummy_vdaf, install_test_trace_subscriber, run_vdaf},
