+++ aggregator/src/aggregator/report_writer.rs
@@ -1,7 +1,7 @@
-use super::{query_type::UploadableQueryType, Error};
-use crate::datastore::{self, models::LeaderStoredReport, Datastore, Transaction};
+use crate::aggregator::{query_type::UploadableQueryType, Error};
 use async_trait::async_trait;
 use futures::future::join_all;
+use janus_aggregator_core::datastore::{self, models::LeaderStoredReport, Datastore, Transaction};
 use janus_core::time::Clock;
 use prio::vdaf;
 use std::{fmt::Debug, marker::PhantomData, mem::replace, sync::Arc, time::Duration};
