+++ aggregator_core/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "janus_aggregator"
+name = "janus_aggregator_core"
 edition.workspace = true
 homepage.workspace = true
 license.workspace = true
@@ -8,22 +8,8 @@ repository.workspace = true
 rust-version.workspace = true
 version.workspace = true
 
-[package.metadata.docs.rs]
-all-features = true
-rustdoc-args = ["--cfg", "docsrs"]
-
 [features]
-default = ["kube-rustls"]
-fpvec_bounded_l2 = ["dep:fixed", "janus_core/fpvec_bounded_l2"]
-tokio-console = ["dep:console-subscriber"]
-jaeger = ["dep:tracing-opentelemetry", "dep:opentelemetry-jaeger"]
-otlp = [
-    "dep:tracing-opentelemetry",
-    "dep:opentelemetry-otlp",
-    "dep:opentelemetry-semantic-conventions",
-    "dep:tonic",
-]
-prometheus = ["dep:opentelemetry-prometheus", "dep:prometheus"]
+default = []
 test-util = [
     "janus_core/test-util",
     "janus_messages/test-util",
@@ -31,31 +17,24 @@ test-util = [
     "dep:lazy_static",
     "dep:testcontainers",
 ]
-kube-rustls = ["kube/rustls-tls"]
-kube-openssl = ["kube/openssl-tls"]
 
 [dependencies]
 async-trait = "0.1"
 anyhow = "1"
-atty = "0.2"
 backoff = { version = "0.4.0", features = ["tokio"] }
 base64 = "0.21.0"
 bytes = "1.4.0"
 chrono = "0.4"
-clap = { version = "4.1.8", features = ["derive", "env"] }
 console-subscriber = { version = "0.1.8", optional = true }
 deadpool = { version = "0.9.5", features = ["rt_tokio_1"] }
 deadpool-postgres = "0.10.5"
 derivative = "2.2.0"
 futures = "0.3.26"
 fixed = { version = "1.23", optional = true }
-git-version = "0.3.5"
 hex = { version = "0.4.3", features = ["serde"], optional = true }
 http = "0.2.9"
 http-api-problem = "0.56.0"
 hyper = "0.14.24"
-itertools = "0.10.5"
-janus_aggregator_core.workspace = true
 janus_core.workspace = true
 janus_messages.workspace = true
 k8s-openapi.workspace = true
@@ -77,8 +56,6 @@ ring = "0.16.20"
 serde = { version = "1.0.152", features = ["derive"] }
 serde_json = "1.0.93"
 serde_yaml = "0.9.17"
-signal-hook = "0.3.15"
-signal-hook-tokio = { version = "0.3.1", features = ["futures-v0_3"] }
 testcontainers = { version = "0.14.0", optional = true }
 thiserror = "1.0"
 tokio = { version = "1.25", features = ["full", "tracing"] }
@@ -87,26 +64,16 @@ tonic = { version = "0.8", optional = true, features = ["tls", "tls-webpki-roots
 tracing = "0.1.37"
 tracing-log = "0.1.3"
 tracing-opentelemetry = { version = "0.18", optional = true }
-tracing-stackdriver = "0.6.2"
 tracing-subscriber = { version = "0.3", features = ["std", "env-filter", "fmt", "json"] }
 url = { version = "2.3.1", features = ["serde"] }
 uuid = { version = "1.3.0", features = ["v4"] }
-warp = { version = "0.3", features = ["tls"] }
 
 [dev-dependencies]
 assert_matches = "1"
 hyper = "0.14.24"
-# Enable `kube`'s `openssl-tls` feature (which takes precedence over the
-# `rustls-tls` feature when creating a default client) to work around rustls's
-# lack of support for connecting to servers by IP addresses, which affects many
-# Kubernetes clusters. Enable the `test-util` feature for various utilities
-# used in unit tests.
-janus_aggregator = { workspace = true, features = ["fpvec_bounded_l2", "kube-openssl", "test-util"] }
-mockito = "0.32.3"
+serde_test = "1.0.152"
 tempfile = "3.4.0"
 tokio = { version = "1", features = ["test-util"] }  # ensure this remains compatible with the non-dev dependency
-trycmd = "0.14.12"
-wait-timeout = "0.2.0"
 
 [build-dependencies]
 rustc_version = "0.4.0"
