+++ core/Cargo.toml
@@ -30,6 +30,7 @@ test-util = [
 ]
 
 [dependencies]
+assert_matches = { version = "1", optional = true }
 backoff = { version = "0.4.0", features = ["tokio"] }
 base64 = "0.21.0"
 chrono = "0.4"
@@ -41,27 +42,24 @@ http-api-problem = "0.56.0"
 janus_messages.workspace = true
 kube = { workspace = true, optional = true, features = ["rustls-tls"] }
 k8s-openapi = { workspace = true, optional = true }
+lazy_static = { version = "1", optional = true }
 prio.workspace = true
 rand = "0.8"
 reqwest = { version = "0.11.14", default-features = false, features = ["rustls-tls", "json"] }
 ring = "0.16.20"
 serde = { version = "1.0.152", features = ["derive"] }
+serde_json = { version = "1.0.93", optional = true }
+serde_yaml = "0.9.17"
+tempfile = { version = "3", optional = true }
+testcontainers = { version = "0.14", optional = true }
 thiserror = "1.0"
 tokio = { version = "1.25", features = ["macros", "net", "rt"] }
 tracing = "0.1.37"
-
-# Dependencies required only if feature "test-util" is enabled
-assert_matches = { version = "1", optional = true }
-lazy_static = { version = "1", optional = true }
-serde_json = { version = "1.0.93", optional = true }
-tempfile = { version = "3", optional = true }
-testcontainers = { version = "0.14", optional = true }
 tracing-log = { version = "0.1.3", optional = true }
 tracing-subscriber = { version = "0.3", features = ["std", "env-filter", "fmt"], optional = true }
 
 [dev-dependencies]
 hex = { version = "0.4", features = ["serde"] }  # ensure this remains compatible with the non-dev dependency
-janus_core = { path = ".", features = ["fpvec_bounded_l2", "test-util"] }
 # Enable `kube`'s `openssl-tls` feature (which takes precedence over the
 # `rustls-tls` feature when creating a default client) to work around rustls's
 # lack of support for connecting to servers by IP addresses, which affects many
