+++ integration_tests/tests/divviup_ts.rs
@@ -1,6 +1,7 @@
 //! These tests check interoperation between the divviup-ts client and Janus aggregators.
 
 use common::{submit_measurements_and_verify_aggregate, test_task_builders};
+use janus_aggregator_core::task::QueryType;
 use janus_core::{
     task::VdafInstance,
     test_util::{install_test_trace_subscriber, testcontainers::container_client},
@@ -16,7 +17,7 @@ mod common;
 
 async fn run_divviup_ts_integration_test(container_client: &Cli, vdaf: VdafInstance) {
     let (collector_private_key, leader_task, helper_task) =
-        test_task_builders(vdaf, janus_aggregator::task::QueryType::TimeInterval);
+        test_task_builders(vdaf, QueryType::TimeInterval);
     let leader_task = leader_task.build();
     let network = generate_network_name();
     let leader = Janus::new_in_container(container_client, &network, &leader_task).await;
