+++ interop_binaries/Cargo.toml
@@ -32,6 +32,7 @@ fixed = { version = "1.23", optional = true }
 fixed-macro = { version = "1.1.1", optional = true }
 hex = { version = "0.4", optional = true }
 lazy_static = { version = "1", optional = true }
+janus_aggregator_core.workspace = true
 janus_aggregator.workspace = true
 janus_client.workspace = true
 janus_collector.workspace = true
