+++ messages/Cargo.toml
@@ -30,8 +30,6 @@ thiserror = "1.0"
 [dev-dependencies]
 assert_matches = "1"
 serde_test = "1.0.152"
-tracing-log = "0.1.3"
-tracing-subscriber = { version = "0.3", features = ["std", "env-filter", "fmt"] }
 trycmd = "0.14.12"
 
 [[bin]]
